Add Rust EventGrid client
