+++ event-models/rust/Cargo.toml
@@ -9,4 +9,5 @@ chrono = { version = "0.4", features = ["serde"] }
 schemafy = "0.5"
 serde = "1"
 serde_json = "1"
-uuid = { version = "0.8.1", features = ["v4"] }
+uuid = { version = "0.8", features = ["v4"] }
+surf = "1"
